I've added 'gulp jsBeautify' to the gulp file.
To check whether you're code is formatted correctly, run 'gulp jsBeautify' from the command line. It will then print out all the
places where javascript doesn't meet our coding style. The style is codified in .jsbeautifyrc. Once we are comfortable with the
style it will then be not only verify the style but write the changes as well.